feat(frontend): add profile dropdown
Changes default page to "aktualności" and adds profile dropdown
