+++ frontend/src/components/Layout.tsx
@@ -1,8 +1,10 @@
-import { chakra, Container, VStack } from "@chakra-ui/react";
+import { Center, chakra, Container, Text, VStack } from "@chakra-ui/react";
 import { Lato } from "@next/font/google";
+import { useSession } from "next-auth/react";
 import type React from "react";
 
 import { Footer } from "./Footer";
+import { Loader } from "./Loader";
 import { Navbar } from "./Navbar";
 
 interface LayoutProps {
@@ -17,11 +19,23 @@ const lato = Lato({
 });
 
 export default function Layout({ children }: LayoutProps) {
+  const { status } = useSession();
+
   return (
     <VStack minH="100vh" spacing={8} className={`${lato.variable}`}>
       <Container maxW="container.xl" flex={1} display="flex" flexDir="column">
         <Navbar />
-        <chakra.main>{children}</chakra.main>
+        <chakra.main>
+          <>
+            {status === "authenticated" ? children : null}
+            {status === "loading" ? <Loader /> : null}
+            {status === "unauthenticated" ? (
+              <Center>
+                <Text>Aby korzystać z aplikacji musisz się zalogować</Text>
+              </Center>
+            ) : null}
+          </>
+        </chakra.main>
       </Container>
       <Footer />
     </VStack>
